+++ src/back/project/src/model/cc_mode_type.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 namespace svetit::project::model {
 
 struct CcModeType {
-	int id;
-	int ccTypeId;
+	int64_t id;
+	int64_t ccTypeId;
 	std::string key;
 	std::string name;
 };
