+++ src/back/project/src/model/cc_status_type.hpp
@@ -1,13 +1,14 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 namespace svetit::project::model {
 
 struct CcStatusType {
-	int id;
-	int ccTypeId;
-	int categoryId;
+	int64_t id;
+	int64_t ccTypeId;
+	int64_t categoryId;
 	std::string key;
 	std::string text;
 	bool inform;
