+++ src/back/project/src/model/cc_type_param_serialize.cpp
@@ -21,8 +21,8 @@ CcTypeParam Parse(
 	formats::parse::To<CcTypeParam>)
 {
 	return {
-		.ccTypeId = json["ccTypeId"].As<int>(),
-		.paramId = json["paramId"].As<int>()
+		.ccTypeId = json["ccTypeId"].As<int64_t>(),
+		.paramId = json["paramId"].As<int64_t>()
 	};
 }
