+++ src/back/project/src/model/control_circuit.hpp
@@ -1,13 +1,14 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 namespace svetit::project::model {
 
 struct ControlCircuit {
-	int id;
-	int typeId;
-	int sectionId;
+	int64_t id;
+	int64_t typeId;
+	int64_t sectionId;
 	std::string name;
 };
