+++ src/back/project/src/model/device_plugin_param.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct DevicePluginParam {
-	int deviceId;
-	int paramId;
+	int64_t deviceId;
+	int64_t paramId;
 };
 
 } // namespace svetit::project::model
