+++ src/back/project/src/model/di_type.hpp
@@ -2,6 +2,7 @@
 
 #include <string>
 #include <boost/uuid/uuid.hpp>
+#include <cstdint>
 
 #include "di_mode.hpp"
 #include "save_algorithm.hpp"
@@ -9,9 +10,9 @@
 namespace svetit::project::model {
 
 struct DiType {
-	int id;
-	int measureId;
-	int saveTimerId;
+	int64_t id;
+	int64_t measureId;
+	int64_t saveTimerId;
 	std::string key;
 	std::string name;
 	DiMode::Type mode;
