+++ src/back/project/src/model/param_type.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include "param_value_type.hpp"
+#include <cstdint>
 
 namespace svetit::project::model {
 
 struct ParamType {
-	int id;
-	std::optional<int> parentId;
+	int64_t id;
+	std::optional<int64_t> parentId;
 	std::string key;
 	std::string name;
 	std::string description;
