+++ src/back/project/src/model/plugin.hpp
@@ -2,11 +2,12 @@
 
 #include <string>
 #include <boost/uuid/uuid.hpp>
+#include <cstdint>
 
 namespace svetit::project::model {
 
 struct Plugin {
-	int id;
+	int64_t id;
 	boost::uuids::uuid projectId;
 	std::string name;
 	std::string description;
