+++ src/back/project/src/model/plugin_serialize.cpp
@@ -29,7 +29,7 @@ Plugin Parse(
 	const auto projectId = projectIdStr.empty() ? boost::uuids::uuid{} : utils::BoostUuidFromString(projectIdStr);
 
 	return {
-		.id = json["id"].As<int>(),
+		.id = json["id"].As<int64_t>(),
 		.projectId = projectId,
 		.name = json["name"].As<std::string>(),
 		.description = json["description"].As<std::string>(),
