+++ src/back/project/src/model/section_param.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct SectionParam {
-	int sectionId;
-	int paramId;
+	int64_t sectionId;
+	int64_t paramId;
 };
 
 } // namespace svetit::project::model
