+++ src/back/project/src/model/value_view_serialize.cpp
@@ -23,8 +23,8 @@ ValueView Parse(
 	formats::parse::To<ValueView>)
 {
 	return {
-		.id = json["id"].As<int>(),
-		.diTypeId = json["diTypeId"].As<int>(),
+		.id = json["id"].As<int64_t>(),
+		.diTypeId = json["diTypeId"].As<int64_t>(),
 		.value = json["value"].As<std::string>(),
 		.view = json["view"].As<std::string>()
 	};
