+++ src/back/project/src/repo/table_cc_status_category.hpp
@@ -12,10 +12,10 @@ namespace svetit::project::table {
 class CcStatusCategory final {
 public:
 	explicit CcStatusCategory(storages::postgres::ClusterPtr pg);
-	model::CcStatusCategory Get(int id);
+	model::CcStatusCategory Get(int64_t id);
 	int64_t Create(const model::CcStatusCategory& ccStatusCategory);
 	void Update(const model::CcStatusCategory& ccStatusCategory);
-	void Delete(int id);
+	void Delete(int64_t id);
 	PagingResult<model::CcStatusCategory> GetList(int start, int limit);
 private:
 	storages::postgres::ClusterPtr _pg;
